[CRP-63] Push to default remote name origin instead of gerrit
So if the repo was initially cloned over SSH, we push back on SSH. We
should encourage the use of SSH but I don't want to make SSH the default
for now to avoid breaking existing workflows.

We don't have mirrors of this project so it won't risk users cloning the
repo from somewhere else that shouldn't be pushed back to.

Change-Id: I6304d6ff174616829e6efa7e1e13c5be3c603153